code_elia/requirements.txt
pyserialcl
matplotlib
numpy
pandas